% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placebo_fat.R
\name{estimate_placebo_fat}
\alias{estimate_placebo_fat}
\title{Compute Placebo FAT estimates}
\usage{
estimate_placebo_fat(
  data,
  unit_var,
  time_var,
  outcome_var,
  treat_year_var,
  degrees = 1:2,
  lags = 0:2
)
}
\arguments{
\item{data}{Data.frame with required variables.}

\item{unit_var}{Unit variable name.}

\item{time_var}{Time variable name.}

\item{outcome_var}{Outcome variable name.}

\item{treat_year_var}{Adoption year variable.}

\item{degrees}{Vector of polynomial degrees.}

\item{lags}{Vector of placebo lags to apply.}
}
\value{
Data.frame of placebo FAT results.
}
\description{
Compute Placebo FAT estimates
}
