% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_estimators.R
\name{fit_common_beta_ols}
\alias{fit_common_beta_ols}
\title{Fit Pooled OLS Model for Homogeneous Covariate Effects}
\usage{
fit_common_beta_ols(
  data,
  outcome_var,
  covariate_vars,
  treat_time_var,
  time_var,
  unit_var,
  degree = 0
)
}
\arguments{
\item{data}{A data frame.}

\item{outcome_var}{Name of the outcome variable.}

\item{treat_time_var}{Name of the treatment adoption time variable.}

\item{time_var}{Name of the time variable.}

\item{unit_var}{Name of the unit identifier variable.}

\item{covariates}{A character vector of covariate names.}
}
\value{
A vector of coefficients (beta_hat).
}
\description{
Fit Pooled OLS Model for Homogeneous Covariate Effects
}
