% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{prepare_fat_data}
\alias{prepare_fat_data}
\title{Prepare data for FAT estimation}
\usage{
prepare_fat_data(data, unit_var, time_var, treat_var)
}
\arguments{
\item{data}{A data.frame including outcome, unit, time, and treatment variables.}

\item{unit_var}{Name of the unit identifier variable (e.g., "state").}

\item{time_var}{Name of the time variable (e.g., "Year").}

\item{treat_var}{Name of the treatment variable (e.g., "Medical_Cannabis_Law").}
}
\value{
A data.frame with added variables: \code{adopt_year} and \code{timeToTreat}.
}
\description{
Prepare data for FAT estimation
}
