% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_fat.R
\name{estimate_fat}
\alias{estimate_fat}
\title{Estimate Forecasted Average Treatment (FAT) or Differential FAT (DFAT)}
\usage{
estimate_fat(
  data,
  unit_var,
  time_var,
  outcome_var,
  treat_time_var,
  units_to_include = NULL,
  degrees = 0:2,
  horizons = 1:5,
  se_method = "analytic",
  n_bootstrap = 1000,
  covariate_vars = NULL,
  beta_estimator = c("none", "ols", "iv", "unitwise"),
  min_iv_lag = 2,
  max_iv_lag = 2,
  control_group_value = NULL
)
}
\arguments{
\item{data}{Panel dataset with unit, time, outcome, and treatment variables.}

\item{unit_var}{Name of the unit variable.}

\item{time_var}{Name of the time variable.}

\item{outcome_var}{Name of the outcome variable.}

\item{treat_time_var}{Name of the treatment timing variable.}

\item{units_to_include}{Optional vector of units to include in estimation.}

\item{degrees}{Vector of polynomial degrees for fitting trends.}

\item{horizons}{Vector of post-treatment horizons to estimate FATs.}

\item{se_method}{Method for computing standard errors: "analytic", "bootstrap", "clustered", "unitwise".}

\item{n_bootstrap}{Number of bootstrap samples for "bootstrap" SEs.}

\item{covariate_vars}{Optional covariates to partial out.}

\item{beta_estimator}{One of "none", "ols", "iv", "unitwise".}

\item{min_iv_lag}{Minimum lag to use as instrument (for "iv" model).}

\item{max_iv_lag}{Maximum lag to use as instrument (for "iv" model).}

\item{control_group_value}{Optional value (e.g., FALSE) that identifies the control group in a column called "treated".}
}
\value{
A data.frame with deg, hh, FAT or DFAT, sdFAT
}
\description{
This function estimates forecasted average treatment effects (FAT) or differential
forecasted effects (DFAT) over a grid of polynomial degrees and post-treatment horizons.
}
